refactor(fe): follow new component directory rule
