+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/problem/_components/SubmissionDetailContent.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import CodeEditor from '@/components/CodeEditor'
-import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area'
+import { ScrollArea, ScrollBar } from '@/components/shadcn/scroll-area'
 import {
   Table,
   TableBody,
@@ -9,7 +9,7 @@ import {
   TableHead,
   TableHeader,
   TableRow
-} from '@/components/ui/table'
+} from '@/components/shadcn/table'
 import { dateFormatter, getResultColor } from '@/lib/utils'
 import type { ContestProblem, Language, SubmissionDetail } from '@/types/type'
